Fix statvfs field types for macOS compatibility
On macOS f_bavail is u32 and f_frsize is u64, unlike Linux where
both are u64. Cast both to u64 before multiplying.

+++ crates/sracha/src/main.rs
@@ -818,7 +818,8 @@ fn available_space(path: &Path) -> Option<u64> {
     if ret != 0 {
         return None;
     }
-    Some(stat.f_bavail * stat.f_frsize)
+    #[allow(clippy::unnecessary_cast)]
+    Some(stat.f_bavail as u64 * stat.f_frsize as u64)
 }
 
 #[cfg(not(unix))]
